+++ src/snowflake/snowpark/_internal/analyzer/schema_utils.py
@@ -86,13 +86,16 @@ def analyze_attributes(
         return get_attributes()
     if lowercase.startswith("describe"):
         session._run_query(sql)
-        return convert_result_meta_to_attribute(session._conn._cursor.description)
+        return convert_result_meta_to_attribute(
+            session._conn._cursor.description, session.structured_types_enabled
+        )
 
     return session._get_result_attributes(sql)
 
 
 def convert_result_meta_to_attribute(
     meta: Union[List[ResultMetadata], List["ResultMetadataV2"]],  # pyright: ignore
+    structured_types_enabled: bool = False,
 ) -> List[Attribute]:
     # ResultMetadataV2 may not currently be a type, depending on the connector
     # version, so the argument types are pyright ignored
@@ -103,7 +106,7 @@ def convert_result_meta_to_attribute(
         attributes.append(
             Attribute(
                 quoted_name,
-                convert_metadata_to_sp_type(column_metadata),
+                convert_metadata_to_sp_type(column_metadata, structured_types_enabled),
                 column_metadata.is_nullable,
             )
         )