+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### New Features
 
+- Added support for structured type schema parsing.
 - Added support for creating vectorized UDTFs with `process` method.
 - Added support for dataframe functions:
   - to_timestamp_ltz